services/google-agent/pyproject.toml
[project]
name = "google-system-design-agent"
version = "0.1.0"
description = "Remote Google system design interview agent exposing deterministic skills."
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.3.0",
    "click>=8.1.8",
    "google-adk>=1.0.0",
    "starlette>=0.37.0",
    "sse-starlette>=2.3.5",
    "uvicorn>=0.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.8.4",
    "pre-commit>=4.0.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100
exclude = [
    ".git",
    ".pytest_cache",
    "__pycache__",
    "*.egg-info",
    ".venv",
    "venv",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
extend-select = ["E402"]
ignore = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
